requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
pydantic==2.11.3
pydantic-settings==2.8.1
motor==3.7.0
pymongo==4.12.0
python-jose==3.4.0
cryptography==44.0.2
python-multipart==0.0.20
email-validator==2.2.0
ruff==0.11.5
pytz==2025.2
python-dotenv==1.1.0
bcrypt==4.3.0
